Updated api with test cases

+++ db/migrate/20210215133833_create_phone_numbers.rb
@@ -1,7 +1,7 @@
 class CreatePhoneNumbers < ActiveRecord::Migration[6.1]
   def change
     create_table :phone_numbers do |t|
-      t.integer :number
+      t.bigint :number
       t.references :account, null: false, foreign_key: true
 
       t.timestamps
